refactor: split read-output's paging/cursor half into read-output-page.ts
The --tab work pushed read-output.ts to 555 lines, over the 500 cap. The
pure paging/shaping/cursor code moves to read-output-page.ts (190 lines);
read-output.ts (419) keeps the read logic and the verb, re-exporting the
moved symbols so the module stays the one import site.
